backend/advisor/quant_models/cashflow_model/requirements.txt
pandas==3.0.0
matplotlib==3.10.8
mesa==3.3.1
jinja2==3.1.6
pyyaml==6.0.3
pydantic==2.12.5